requirements.txt
decorator==5.0.9
Django==2.2.10
djangorestframework==3.12.4
pytz==2021.1
self==2020.12.3
sqlparse==0.4.1
